+++ perla-verde-traveler/src/main.jsx
@@ -1,10 +1,11 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter } from "react-router-dom";
+
 import './index.css'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>
+  <BrowserRouter future={{ v7_relativeSplatPath: true }}>
     <App />
-  </StrictMode>,
+  </BrowserRouter>,
 )
