+++ perla-verde-traveler/src/not_found.jsx
@@ -1,0 +1,10 @@
+export default function NotFound() {
+    return (
+        <div className="min-h-screen bg-green-50">
+            <h1 className="text-3xl font-bold text-center mt-12">No se encuentra la ruta</h1>
+            <p className="text-center mt-4">
+                404.
+            </p>
+        </div>
+    );
+}
