+++ perla-verde-traveler/src/pages/attractions.jsx
@@ -1,0 +1,13 @@
+import Header from "../components/header"
+
+export default function Attractions() {
+    return (
+        <div className="min-h-screen bg-green-50">
+            <Header />
+            <h1 className="text-3xl font-bold text-center mt-12">Atracciones</h1>
+            <p className="text-center mt-4">
+                Descubre las atracciones más impresionantes de San Luis.
+            </p>
+        </div>
+    );
+}
